# api.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Vytals
servers:
  - url: 'http://localhost:5000'
paths:
  /users/new:
    post:
      summary: 'Create new user.'
      requestBody:
          description: 'User object will be used to store new user in the system.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      responses:
        '200':
          description: 'User was succesfully stored in the system.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    User:
      type: object
      required:
        - first_name
        - last_name
        - birthdate
        - email
      properties:
        id:
          type: integer
          format: int64
        first_name:
          type: string
        last_name:
          type: string
        birthdate:
          type: string
          format: date-time
        email:
          type: string
          format: email
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string